openapi/requirements-prod.txt
oss2==2.13.1
Django==3.2.2
djangorestframework==3.12.4
mysqlclient
uwsgi==2.0.19.1
requests==2.25.1
https://huoqi-public.oss-cn-beijing.aliyuncs.com/iast/dongtai-latest.tar.gz
